refactor: movefiles into correct module
