+++ packages/shared/lib/contexts/onboarding/actions/backupInitialStronghold.ts
@@ -1,8 +1,7 @@
 import { get } from 'svelte/store'
 
-import { strongholdPassword } from '@contexts/onboarding'
+import { strongholdPassword, updateNewProfile } from '@contexts/onboarding'
 import { Platform } from '@lib/platform'
-import { updateNewProfile } from '@core/profile'
 import { backup } from '@core/profile-manager'
 import { getDefaultStrongholdName } from '@lib/utils'
