+++ packages/shared/lib/contexts/onboarding/actions/cleanupOnboarding.ts
@@ -1,4 +1,4 @@
-import { deleteNewProfile } from '@core/profile/actions/new-profile'
+import { deleteNewProfile } from '@contexts/onboarding/actions/newProfile'
 
 import { cleanupOnboardingStores } from './cleanupOnboardingStores'
