+++ packages/shared/lib/contexts/onboarding/actions/index.ts
@@ -1,7 +1,9 @@
+export * from './addNode'
 export * from './backupInitialStronghold'
 export * from './cleanupOnboarding'
 export * from './cleanupOnboardingStores'
 export * from './cleanupProtectionOnboarding'
+export * from './newProfile'
 export * from './resetImportState'
 export * from './restoreBackupFromFile'
 export * from './setNewProfileClientOptions'
