+++ packages/shared/lib/contexts/onboarding/actions/restoreBackupFromFile.ts
@@ -1,5 +1,6 @@
 import { get } from 'svelte/store'
-import { newProfile, ProfileImportType } from '@core/profile'
+import { ProfileImportType } from '@core/profile'
+import { newProfile } from '@contexts/onboarding'
 import { importFilePath, importType, isGettingMigrationData } from '../stores'
 import { Platform } from '@lib/platform'
 import { getMigrationData } from '@lib/migration'
