+++ packages/shared/lib/contexts/onboarding/stores/index.ts
@@ -2,4 +2,5 @@ export * from './import-file-path.store'
 export * from './import-type.store'
 export * from './is-getting-migration-data.store'
 export * from './mnemonic.store'
+export * from './new-profile.store'
 export * from './stronghold-password.store'
