+++ packages/shared/lib/core/profile/actions/active-profile/addNodeToActiveProfile.ts
@@ -1,6 +1,6 @@
 import { getNodeInfo, setClientOptions } from '@core/profile-manager'
 import { get } from 'svelte/store'
-import { checkNodeUrlValidity, INode, updateNodeInfo, buildNode, nodeInfo } from '@core/network'
+import { INode, updateNodeInfo, buildNode, nodeInfo } from '@core/network'
 import { activeProfile, updateActiveProfile } from '@core/profile/stores'
 
 export async function addNodeToActiveProfile(node: INode): Promise<void> {
@@ -11,15 +11,9 @@ export async function addNodeToActiveProfile(node: INode): Promise<void> {
     updateNodeInfo(nodeInfoResponse?.nodeInfo)
     const builtNode = buildNode(nodeInfoResponse, node.auth)
 
-    const hasValidUrl =
-        checkNodeUrlValidity(
-            _activeProfile.clientOptions?.nodes,
-            builtNode?.url,
-            _activeProfile?.isDeveloperProfile
-        ) === undefined
     const isInSameNetwork = oldNodeInfo.protocol.networkName === nodeInfoResponse.nodeInfo.protocol.networkName
 
-    if (hasValidUrl && isInSameNetwork) {
+    if (isInSameNetwork) {
         const nodes = [..._activeProfile?.clientOptions?.nodes, builtNode]
         const clientOptions = {
             nodes,
