+++ packages/shared/lib/core/profile/actions/index.ts
@@ -1,4 +1,3 @@
 export * from './active-accounts'
 export * from './active-profile'
-export * from './new-profile'
 export * from './profiles'
