+++ packages/shared/lib/core/profile/stores/index.ts
@@ -3,7 +3,6 @@ export * from './active-profile.store'
 export * from './active-accounts.store'
 export * from './is-software-profile.store'
 export * from './is-ledger-profile.store'
-export * from './new-profile.store'
 export * from './profiles.store'
 export * from './timeSinceStrongholdUnlocked.store'
 export * from './timeStrongholdLastUnlocked.store'
