+++ packages/shared/lib/core/router/app-router.ts
@@ -2,15 +2,8 @@ import { get, writable } from 'svelte/store'
 
 import { AppStage, appStage, mobile } from '@core/app'
 import { NetworkType } from '@core/network'
-import {
-    activeProfile,
-    newProfile,
-    ProfileImportType,
-    profiles,
-    ProfileType,
-    setNewProfileType,
-    updateNewProfile,
-} from '@core/profile'
+import { activeProfile, ProfileImportType, profiles, ProfileType } from '@core/profile'
+import { newProfile, setNewProfileType, updateNewProfile } from '@contexts/onboarding'
 import { SetupType } from '@lib/typings/setup'
 import { walletSetupType } from '@lib/wallet'
