+++ packages/shared/routes/onboarding/CreateView.svelte
@@ -3,7 +3,8 @@
     import features from 'shared/features/features'
     import { mobile } from '@core/app'
     import { localize } from '@core/i18n'
-    import { ProfileType, setNewProfileType, newProfile } from '@core/profile'
+    import { ProfileType, newProfile } from '@core/profile'
+    import { setNewProfileType } from '@contexts/onboarding'
     import { appRouter } from '@core/router'
 
     function handleContinueClick(profileType: ProfileType): void {
