+++ packages/shared/routes/onboarding/PasswordView.svelte
@@ -53,7 +53,6 @@
                 $strongholdPassword = password
                 $appRouter.next({ password })
             } catch (err) {
-                console.error(err)
                 showAppNotification({
                     type: 'error',
                     message: localize(err.error),
