+++ packages/shared/routes/onboarding/import/views/ImportView.svelte
@@ -4,7 +4,8 @@
     import features from 'shared/features/features'
     import { mobile } from '@core/app'
     import { localize } from '@core/i18n'
-    import { newProfile, ProfileImportType, ProfileType, setNewProfileType } from '@core/profile'
+    import { newProfile, ProfileImportType, ProfileType } from '@core/profile'
+    import { setNewProfileType } from '@contexts/onboarding'
 
     const dispatch = createEventDispatcher()
